--- components/common/modal/modalTemplate/ModalTemplate.type.ts
@@ -1,3 +1,0 @@
-export interface ModalTemplatePropsType {
-    hide: (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-}
