+++ components/initialRegister/button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from './Button';
+import Button from '.';
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default {
