+++ components/initialRegister/setUserInterest/index.tsx
