+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import Modal from '../components/common/modal/Modal'
-import ModalTemplate from '../components/common/modal/modalTemplate/ModalTemplate';
-import SetUserProfile from '../components/initialRegister/setUserProfile/SetUserProfile';
+import ModalTemplate from '../components/common/modal/ModalTemplate';
+import SetUserProfile from '../components/initialRegister/SetUserProfile';
 
 import useModal from '../hooks/useModal';
